Rework cooldowns for Rik Reverb
Looking at the damage intake on this fight I wanted to try and
plan only a handful of cooldown and then the team can sort of feel it
out. I think the biggest thing will be trying to not overlap healing
cooldowns but to use them as mana saving buttons. This schedules Invoke
Yu'lon and some DR cooldowns; Yu'lon mostly because it lasts a while and
it will cover the whole time where we aren't draining the amplifiers
anymore. Aura Mastery, Zephyr and Rally in those spots will also help a
bit.